refactor: add function division in report functions
